+++ xyz.meedu.api/app/Console/Commands/MeEduUpgradeCommand.php
@@ -21,7 +21,12 @@ class MeEduUpgradeCommand extends Command
 
     public function handle()
     {
-        // JWT 密钥校验
+        // 敏感密钥校验
+        if (empty(config('app.key'))) {
+            $this->error('APP_KEY 未配置!请在 .env 中设置后再执行升级。');
+            $this->warn('可执行 `php artisan key:generate` 自动生成并写入 .env');
+            return 1;
+        }
         if (empty(config('jwt.secret'))) {
             $this->error('JWT_SECRET 未配置!请在 .env 中设置后再执行升级。');
             $this->warn('可执行 `php artisan jwt:secret` 自动生成并写入 .env');